requirements.txt
python-dotenv
flet
requests